Multi monitor support
Was testing this and it took my game from my larger monitor that I play on and moved it to my primary monitor.  This change checks the monitor that the game is currently running on and fullscreens it there instead of defaulting to 0,0 which is top left of the system's primary monitor.