Version 7.1
Features added:

* Added an option to adjust the sizing region for windows (this resolves
[issue #44](https://github.com/Codeusa/Borderless-Gaming/issues/44),
[issue #53](https://github.com/Codeusa/Borderless-Gaming/issues/53),
[issue #53](https://github.com/Codeusa/Borderless-Gaming/issues/56),
[issue #71](https://github.com/Codeusa/Borderless-Gaming/issues/71), and
[issue #84](https://github.com/Codeusa/Borderless-Gaming/issues/84)).
For example, with PAYDAY 2, make sure the option to auto-maximize the
window is unchecked and set the bounding to -2, 0, -2, 2.
* Added an option to close-to-tray (this resolves [issue #60 part
1](https://github.com/Codeusa/Borderless-Gaming/issues/60)).
* Added an option to hide the tray balloon tips (tray bubble
notifications) (this resolves [issue #60 part
2](https://github.com/Codeusa/Borderless-Gaming/issues/60)).
* Added an option to start the program minimized (this resolves [issue
#61](https://github.com/Codeusa/Borderless-Gaming/issues/61)).
* Added an option to not auto-maximize windows when they are made
borderless.  Auto-maximizing windows was added in version 7.0, but
conflicts with specific bounding options for games that don't play
nicely.
* Added a way to restore a window to its previous un-borderless windowed
state.  For this to work, the window must have been made borderless by
Borderless Gaming (and in this session: you can't exit Borderless Gaming
or it will forget the original bordered data).  Additionally, any menus
the window may have had will be lost.  It's recommended to exit and
re-launch any applications that you want restored to a bordered state,
but this feature may act as a backup in cases where that's inconvenient.

Miscellaneous:

* New method of determining internet availability added.
* Updated credits.
* Minor typos and grammatical errors fixed.
* Copyright updated to 2014-2015.
* Minor code cleanup and refactoring.