Version 8.3
Changes:
Added exception handling when reading Favorites settings so that in the
case of corrupt or unreadable favorites, the application may continue to
load without them.  This resolves issue #191.
Added more exception handling when enumerating processes to prevent
issues where Windows is resuming from sleep or hibernation and a process
is not responding or unavailable in an unusual way.  This resolves issue
#208.
Merged changes from SecretOnline, including adding a process to
favorites by regex string and togglable application update availability
check.
Updated the program icon with a new font that's more visible at more
resolutions.