Version 8.4
This is a minor update that swaps the Aurora Group copyright text and
image from the About menu and replaces it with Codeusa Software again.
Fixed a minor menu issues with the Favorites list right-click showing
checkboxes at odd offsets.  Added a statusbar label hint for
right-clicking since it's not obvious you can do that and wasn't present
in pre-6.5 versions.  Added a menu option to the Help menu that links to
the 'How to use' guide at the Steam discussion board.  Updated the
solution and installer to multiple projects that compile to different
binaries with individual projects overlapping source files wherever
possible.