Disable Modal for Rainway
On a more personal note I am deeply saddened by the response to this. While it should have been implemented better, the outrage and misinformation that resulted from it really hit me hard.

I've always kept Borderless Gaming close to my chest and wanted it to be one of the best gaming utils out there. I helped create another passion project and wanted to promote it through an existing channel -- you build software and give it away for free and you are some how the worst person ever.